Rename class GroupForward -> GroupForwardTest, Add 2 tests to GroupForwardTest
