--- src/test/java/school/redrover/GroupForward.java
@@ -1,5 +1,0 @@
-package school.redrover;
-
-public class GroupForward {
-
-}
